--- selfdrive/ui/tests/diff/replay_setup.py
@@ -1,97 +1,0 @@
-from __future__ import annotations
-from typing import TYPE_CHECKING
-from cereal import car, log, messaging
-from collections.abc import Callable
-
-from cereal.messaging import PubMaster
-from openpilot.common.basedir import BASEDIR
-from openpilot.common.params import Params
-from openpilot.selfdrive.selfdrived.alertmanager import set_offroad_alert
-from openpilot.system.updated.updated import parse_release_notes
-from openpilot.system.version import terms_version, training_version
-
-if TYPE_CHECKING:
-  # Prevent circular import for type checking
-  from openpilot.selfdrive.ui.tests.diff.replay import ReplayContext
-
-BRANCH_NAME = "this-is-a-really-super-mega-ultra-max-extreme-ultimate-long-branch-name"
-
-
-def initialize_params() -> None:
-  """Initialize Params with necessary values for the replay script."""
-  params = Params()
-  params.put("HasAcceptedTerms", terms_version)
-  params.put("CompletedTrainingVersion", training_version)
-  params.put("DongleId", "test123456789")
-  params.put("UpdaterCurrentDescription", "0.10.1 / test-branch / abc1234 / Nov 30")
-
-
-def put_update_params(params: Params | None = None) -> None:
-  if params is None:
-    params = Params()
-  params.put("UpdaterCurrentReleaseNotes", parse_release_notes(BASEDIR))
-  params.put("UpdaterNewReleaseNotes", parse_release_notes(BASEDIR))
-  params.put("UpdaterTargetBranch", BRANCH_NAME)
-
-
-def setup_offroad_alerts() -> None:
-  put_update_params(Params())
-  set_offroad_alert("Offroad_TemperatureTooHigh", True, extra_text='99C')
-  set_offroad_alert("Offroad_ExcessiveActuation", True, extra_text='longitudinal')
-  set_offroad_alert("Offroad_IsTakingSnapshot", True)
-
-
-def setup_update_available() -> None:
-  params = Params()
-  params.put_bool("UpdateAvailable", True)
-  params.put("UpdaterNewDescription", f"0.10.2 / {BRANCH_NAME} / 0a1b2c3 / Jan 01")
-  put_update_params(params)
-
-
-def setup_developer_params() -> None:
-  CP = car.CarParams()
-  CP.alphaLongitudinalAvailable = True
-  Params().put("CarParamsPersistent", CP.to_bytes())
-
-
-def send_onroad(pm: PubMaster) -> None:
-  ds = messaging.new_message('deviceState')
-  ds.deviceState.started = True
-  ds.deviceState.networkType = log.DeviceState.NetworkType.wifi
-
-  ps = messaging.new_message('pandaStates', 1)
-  ps.pandaStates[0].pandaType = log.PandaState.PandaType.dos
-  ps.pandaStates[0].ignitionLine = True
-
-  pm.send('deviceState', ds)
-  pm.send('pandaStates', ps)
-
-
-def make_network_state_setup(ctx: ReplayContext, network_type) -> Callable:
-  def _send() -> None:
-    ds = messaging.new_message('deviceState')
-    ds.deviceState.networkType = network_type
-    ctx.pm.send('deviceState', ds)
-
-  return lambda: ctx.update_send_fn(_send)
-
-
-def make_onroad_setup(ctx: ReplayContext) -> Callable:
-  def _send() -> None:
-    send_onroad(ctx.pm)
-
-  return lambda: ctx.update_send_fn(_send)
-
-
-def make_alert_setup(ctx: ReplayContext, size, text1, text2, status) -> Callable:
-  def _send() -> None:
-    send_onroad(ctx.pm)
-    alert = messaging.new_message('selfdriveState')
-    ss = alert.selfdriveState
-    ss.alertSize = size
-    ss.alertText1 = text1
-    ss.alertText2 = text2
-    ss.alertStatus = status
-    ctx.pm.send('selfdriveState', alert)
-
-  return lambda: ctx.update_send_fn(_send)
